+++ src/main/webapp/WEB-INF/good.jsp
@@ -12,6 +12,6 @@
 </head>
 <body>
 good
-<form action="/selectMember"><button>显示</button></form>
+<form action="/user/selectMember"><button>显示</button></form>
 </body>
 </html>
